FIX: File was not found when it was older than the timestamp of last run
The timestamp was set to utcnow when the run was finished, this unfortunately
skipped any file which were uploaded after that timestamp and was older then the
timestamp. The last_modified time of the file is really last modified time of
the file and not the nextcloud entry. So uploading of the file to server doesn't
change this file property.

To prevent skipping new entry the timestamp is now set to newest last_modified
time found during check of the folders. This way we don't skip any file newer
file regardless of the time of upload to server. If the file is newer than the
timestamp in timestamp_file it will be found.